requirements.txt
SQLAlchemy==1.4.28
Flask==2.0.2
python-decouple==3.5
Flask-SQLAlchemy==2.5.1
psycopg==3.0.5
psycopg2-binary==2.9.2
